deployment/requirements.txt
aws-cdk-lib
constructs>=10.0.0,<11.0.0
sagemaker
requests==2.28.2
streamlit
pandas